requirements.txt
Flask==2.1.1
PyYAML==6.0
openapi-spec-validator==0.7.6
